chore: mv sync fork code to types module
